moved the PromptIOBox to ContentPanelPrompts component because it's not reusable in whole application
