+++ frontend/src/components/ContentPanel/ContentPanelPrompts/PromptIOBox/IOBoxBar/index.tsx
